+++ addons/odoo_atlas/models/__init__.py
@@ -1,0 +1,4 @@
+from . import atlas_conversation
+from . import atlas_message
+from . import atlas_message_citation
+from . import res_config_settings