+++ addons/odoo_atlas/tests/__init__.py
@@ -1,0 +1,4 @@
+from . import common
+from . import test_conversation
+from . import test_citation
+from . import test_security